+++ examples/resources/harness_platform_connector_gcp_cloud_cost/import.sh
@@ -1,0 +1,2 @@
+# Import using gcp cloud cost connector id
+terraform import harness_platform_connector_gcp_cloud_cost.example <connector_id>