+++ internal/service/cd/application/application_gitsync.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/harness/harness-go-sdk/harness/cd/graphql"
 	"github.com/harness/terraform-provider-harness/internal"
+	"github.com/harness/terraform-provider-harness/internal/utils"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -59,7 +60,9 @@ func ResourceApplicationGitSync() *schema.Resource {
 
 func resourceApplicationGitSyncCreateOrUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c := meta.(*internal.Session).CDClient
-
+	if c == nil {
+		return diag.Errorf(utils.CDClientAPIKeyError)
+	}
 	input := &graphql.UpdateApplicationGitSyncConfigInput{}
 
 	if attr := d.Get("branch").(string); attr != "" {
@@ -110,7 +113,9 @@ func readGitSyncConfig(d *schema.ResourceData, config *graphql.GitSyncConfig) {
 
 func resourceApplicationGitSyncRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c := meta.(*internal.Session).CDClient
-
+	if c == nil {
+		return diag.Errorf(utils.CDClientAPIKeyError)
+	}
 	appId := d.Get("app_id").(string)
 
 	app, err := c.ApplicationClient.GetApplicationById(appId)
@@ -132,7 +137,9 @@ func resourceApplicationGitSyncRead(ctx context.Context, d *schema.ResourceData,
 
 func resourceApplicationGitSyncDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c := meta.(*internal.Session).CDClient
-
+	if c == nil {
+		return diag.Errorf(utils.CDClientAPIKeyError)
+	}
 	if err := c.ApplicationClient.RemoveGitSyncConfig(d.Get("app_id").(string)); err != nil {
 		return diag.FromErr(err)
 	}