+++ internal/service/cd/delegate/delegate_approval.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/harness/harness-go-sdk/harness/cd/graphql"
 	"github.com/harness/terraform-provider-harness/internal"
+	"github.com/harness/terraform-provider-harness/internal/utils"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
@@ -46,6 +47,9 @@ func ResourceDelegateApproval() *schema.Resource {
 
 func resourceDelegateApprovalRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c := meta.(*internal.Session).CDClient
+	if c == nil {
+		return diag.Errorf(utils.CDClientAPIKeyError)
+	}
 
 	id := d.Get("delegate_id").(string)
 	delegate, err := c.DelegateClient.GetDelegateById(id)
@@ -74,7 +78,9 @@ func resourceDelegateApprovalRead(ctx context.Context, d *schema.ResourceData, m
 
 func resourceDelegateApprovalCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	c := meta.(*internal.Session).CDClient
-
+	if c == nil {
+		return diag.Errorf(utils.CDClientAPIKeyError)
+	}
 	id := d.Get("delegate_id").(string)
 	delegate, err := c.DelegateClient.GetDelegateById(id)
 	if err != nil {