+++ internal/service/platform/connector/docker.go
@@ -28,12 +28,12 @@ func ResourceConnectorDocker() *schema.Resource {
 				Required:    true,
 			},
 			"url": {
-				Description: "The url of the docker registry.",
+				Description: "The URL of the docker registry.",
 				Type:        schema.TypeString,
 				Required:    true,
 			},
 			"delegate_selectors": {
-				Description: "Connect using only the delegates which have these tags.",
+				Description: "Tags to filter delegates for connection.",
 				Type:        schema.TypeSet,
 				Optional:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
@@ -56,7 +56,7 @@ func ResourceConnectorDocker() *schema.Resource {
 							},
 						},
 						"username_ref": {
-							Description:   "The reference to the username to use for the docker registry.",
+							Description:   "The reference to the Harness secret containing the username to use for the docker registry." + secret_ref_text,
 							Type:          schema.TypeString,
 							Optional:      true,
 							ConflictsWith: []string{"credentials.0.username"},
@@ -66,7 +66,7 @@ func ResourceConnectorDocker() *schema.Resource {
 							},
 						},
 						"password_ref": {
-							Description: "The reference to the password to use for the docker registry.",
+							Description: "The reference to the Harness secret containing the password to use for the docker registry." + secret_ref_text,
 							Type:        schema.TypeString,
 							Required:    true,
 						},