+++ internal/service/platform/connector/prometheus_data_source.go
@@ -12,12 +12,12 @@ func DatasourceConnectorPrometheus() *schema.Resource {
 
 		Schema: map[string]*schema.Schema{
 			"url": {
-				Description: "Url of the Prometheus server.",
+				Description: "URL of the Prometheus server.",
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
 			"delegate_selectors": {
-				Description: "Connect using only the delegates which have these tags.",
+				Description: "Tags to filter delegates for connection.",
 				Type:        schema.TypeSet,
 				Computed:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
@@ -28,7 +28,7 @@ func DatasourceConnectorPrometheus() *schema.Resource {
 				Computed:    true,
 			},
 			"password_ref": {
-				Description: "Password reference.",
+				Description: "Reference to the Harness secret containing the password." + secret_ref_text,
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
@@ -44,7 +44,7 @@ func DatasourceConnectorPrometheus() *schema.Resource {
 							Computed:    true,
 						},
 						"encrypted_value_ref": {
-							Description: "Encrypted value reference.",
+							Description: "Reference to the Harness secret containing the encrypted value." + secret_ref_text,
 							Type:        schema.TypeString,
 							Computed:    true,
 						},